+++ LambdaCore/Core/FunctionExecutor.cs
@@ -19,7 +19,7 @@ namespace LambdaManager.Core
 
         private static readonly Location OutputAlloc = new Location();
 
-        public static DataType.Solution Solution { get; set; } = new DataType.Solution();
+        public static Solution Solution { get; set; } = new Solution();
 
 
         public static int Evaluate(ExecInfo info)